requirements.txt
fastapi
uvicorn[standard]
pillow
requests
paddleocr[doc-parser]
fastapi-queue
redis
python-dotenv
